Module ledger_mob_apdu::digest

Helpers for computing APDU / event digests

This is used in place of Digestible as the same digest must be computed over engine events and APDUs.

Functions

digest_ring_add_txout

digest_ring_init

digest_ring_set_blinding

digest_ring_sign

digest_tx_init

digest_tx_set_message

digest_tx_sign_memo

digest_tx_summary_add_input

digest_tx_summary_add_output

digest_tx_summary_add_output_unblinding

digest_tx_summary_build

digest_tx_summary_init