Module ledger_mob_apdu::prelude

Prelude to simplify downstream use of APDU objects

Re-exports

pub use crate::app_info::AppFlags;

pub use crate::app_info::AppInfoReq;

pub use crate::app_info::AppInfoResp;

pub use crate::ident::IdentGetReq;

pub use crate::ident::IdentResp;

pub use crate::ident::IdentSignReq;

pub use crate::key_image::KeyImageReq;

pub use crate::key_image::KeyImageResp;

pub use crate::random::RandomReq;

pub use crate::random::RandomResp;

pub use crate::subaddress_keys::SubaddressKeyReq;

pub use crate::subaddress_keys::SubaddressKeyResp;

pub use crate::tx::TxComplete;

pub use crate::tx::TxInfo;

pub use crate::tx::TxInfoReq;

pub use crate::wallet_keys::WalletKeyReq;

pub use crate::wallet_keys::WalletKeyResp;

Structs

AddTxOutFlags

TxSummaryAddTxOut flags

TxAddTxOut

Add a TxOut to a ring signing operation

TxGetKeyImage

Fetch a key image for a signed ring, returns TxKeyImage on success (0-byte APDU)

TxGetResponse

Fetch a response scalar for a ring entry in a signed ring, returns TxResponse on success

TxInit

Transaction initialisation APDU, sets up a transaction for execution

TxKeyImage

Key image response APDU for a signed ring

TxMemoSign

Memo HMAC signing request

TxRingInit

Start a ring signing operation

TxRingSign

Execute signing operation for a completed ring (0 length APDU)

TxSetBlinding

Set blinding for ring signing

TxSetMessage

Set the message for the transaction

TxSummaryAddTxIn

Add TxInSummary for a transaction

TxSummaryAddTxOut

Add TxOutSummary to the summary

TxSummaryAddTxOutUnblinding

Add TxOutSummaryUnblinding to the summary

TxSummaryBuild

Complete TxSummary building

TxSummaryInit

Set TxSummary for a transaction

Type Definitions

TxPrivateKey

Transaction private key type (to be moved to mc_core_types)