Module ledger_mob_apdu::tx

Transaction related APDUs, used to execute a transaction via the hardware wallet.

See [ledger_mob_core::engine] for interaction and state machines

Structs

AddTxInFlags

TxSummaryAddTxOut flags

AddTxOutFlags

TxSummaryAddTxOut flags

AddTxOutUnblindingFlags

TxSummaryAddTxOut flags

Tx

Marker trait for transaction key type (to be moved to mc_core_types)

TxAddTxOut

Add a TxOut to a ring signing operation

TxComplete

Complete transaction operation (0 length APDU)

TxGetKeyImage

Fetch a key image for a signed ring, returns TxKeyImage on success (0-byte APDU)

TxGetResponse

Fetch a response scalar for a ring entry in a signed ring, returns TxResponse on success

TxHeader

Header shared between TX response APDUs

TxInfo

Transaction information response APDU.

TxInfoReq

Transaction information request APDU

TxInit

Transaction initialisation APDU, sets up a transaction for execution

TxKeyImage

Key image response APDU for a signed ring

TxMemoSig

Memo signature response

TxMemoSign

Memo HMAC signing request

TxOnetime

Marker trait for onetime key type (to be moved to mc_core_types)

TxResponse

TX Response APDU, contains a response scalar for a given entry in the signed ring.

TxRingInit

Start a ring signing operation

TxRingInitFlags

TxRingInit flags

TxRingSign

Execute signing operation for a completed ring (0 length APDU)

TxSetBlinding

Set blinding for ring signing

TxSetMessage

Set the message for the transaction

TxSummaryAddTxIn

Add TxInSummary for a transaction

TxSummaryAddTxOut

Add TxOutSummary to the summary

TxSummaryAddTxOutUnblinding

Add TxOutSummaryUnblinding to the summary

TxSummaryBuild

Complete TxSummary building

TxSummaryInit

Set TxSummary for a transaction

Enums

FogId

Fog identifier for resolving account information

Constants

FOG_IDS

TODO: support old signal fog URLs List of supported fogs for iteration and display

Type Aliases

TxOnetimeKey

Transaction onetime key type (to be moved to mc_core_types)

TxPrivateKey

Transaction private key type (to be moved to mc_core_types)

TxPublicKey

Transaction public key type (to be moved to mc_core_types)